backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pymongo==4.6.0
motor==3.3.1
pydantic==2.4.2
python-dotenv==1.0.0
PyJWT==2.8.0
certifi==2023.7.22
pyOpenSSL==23.3.0
cryptography==41.0.5
dnspython==2.4.2
urllib3==2.1.0
requests==2.31.0